Move serializable tests to own file
